[MetaAddress] Fix tie type
The tuple returned by the `tie` method used to have the type of the
`Type` field (`uint16_t`) for the `Address` field (`uint64_t`) and
viceversa.

+++ include/revng/Support/MetaAddress.h
@@ -796,7 +796,7 @@ private:
   }
 
 private:
-  using Tied = std::tuple<uint32_t, uint16_t, uint64_t, uint16_t>;
+  using Tied = std::tuple<uint32_t, uint16_t, uint16_t, uint64_t>;
   Tied tie() const { return std::tie(Epoch, AddressSpace, Type, Address); }
 };
